Instagram: The ID of an Instagram location should now be represented by Int64 instead of Int32
It appears that the location IDs have now exceeded what is supported by Int32, so we instead need to use Int64.

+++ src/Skybrud.Social/Instagram/Endpoints/InstagramLocationsEndpoint.cs
@@ -37,7 +37,7 @@ namespace Skybrud.Social.Instagram.Endpoints {
         /// </summary>
         /// <param name="locationId">The ID of the location.</param>
         /// <see cref="http://instagram.com/developer/endpoints/locations/#get_locations"/>
-        public InstagramLocationResponse GetLocation(int locationId) {
+        public InstagramLocationResponse GetLocation(long locationId) {
             return InstagramLocationResponse.ParseResponse(Raw.GetLocation(locationId));
         }
         
@@ -67,7 +67,7 @@ namespace Skybrud.Social.Instagram.Endpoints {
         /// </summary>
         /// <param name="locationId">The ID of the location.</param>
         /// <see cref="http://instagram.com/developer/endpoints/locations/#get_locations_media_recent"/>
-        public InstagramLocationRecentMediaResponse GetRecentMedia(int locationId) {
+        public InstagramLocationRecentMediaResponse GetRecentMedia(long locationId) {
             return InstagramLocationRecentMediaResponse.ParseResponse(Raw.GetRecentMedia(locationId));
         }
 
@@ -77,7 +77,7 @@ namespace Skybrud.Social.Instagram.Endpoints {
         /// <param name="locationId">The ID of the location.</param>
         /// <param name="options">The options for the search.</param>
         /// <see cref="http://instagram.com/developer/endpoints/locations/#get_locations_media_recent"/>
-        public InstagramLocationRecentMediaResponse GetRecentMedia(int locationId, InstagramLocationRecentMediaOptions options) {
+        public InstagramLocationRecentMediaResponse GetRecentMedia(long locationId, InstagramLocationRecentMediaOptions options) {
             return InstagramLocationRecentMediaResponse.ParseResponse(Raw.GetRecentMedia(locationId, options));
         }
